article search adn order

+++ article/views.py
@@ -12,15 +12,47 @@ from .forms import ArticlePostForm
 # 引入markdown模块
 import markdown
 from django.contrib.auth.decorators import login_required
-
+# 引入分页模块
+from django.core.paginator import Paginator
+# 引入 Q 对象
+from django.db.models import Q
 # 视图函数
 
 
 def article_list(request):
-    # 取出所有博客文章
-    articles = ArticlePost.objects.all()
+    search = request.GET.get('search')
+    order = request.GET.get('order')
+    # 用户搜索逻辑
+    if search:
+        if order == 'total_views':
+            # 用 Q对象 进行联合搜索
+            article_list = ArticlePost.objects.filter(
+                Q(title__icontains=search) |
+                Q(body__icontains=search)
+            ).order_by('-total_views')
+        else:
+            article_list = ArticlePost.objects.filter(
+                Q(title__icontains=search) |
+                Q(body__icontains=search)
+            )
+    else:
+        # 将 search 参数重置为空
+        search = ''
+        if order == 'total_views':
+            article_list = ArticlePost.objects.all().order_by('-total_views')
+        else:
+            article_list = ArticlePost.objects.all()
+
+    # 分页功能
+    # 每页显示 1 篇文章
+    paginator = Paginator(article_list, 6)
+    # 获取 url 中的页码
+    page = request.GET.get('page')
+    # 将导航对象相应的页码内容返回给 articles
+    articles = paginator.get_page(page)
+
     # 需要传递给模板（templates）的对象
-    context = {'articles': articles}
+    context = {'articles': articles, 'order': order, 'search': search}
     # render函数：载入模板，并返回context对象
     return render(request, 'article/list.html', context)
 
@@ -38,6 +70,8 @@ def article_detail(request, id):
             # 语法高亮扩展
             'markdown.extensions.codehilite',
             ])
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
     # 需要传递给模板的对象
     context = {'article': article}
     # 载入模板，并返回context对象
@@ -76,9 +110,12 @@ def article_create(request):
         return render(request, 'article/create.html', context)
 
 # 删文章
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
     # 根据 id 获取需要删除的文章
     article = ArticlePost.objects.get(id=id)
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     # 调用.delete()方法删除文章
     article.delete()
     # 完成删除后返回文章列表
@@ -94,6 +131,7 @@ def article_safe_delete(request, id):
         return HttpResponse("仅允许post请求")
 
 # 更新文章
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -104,6 +142,8 @@ def article_update(request, id):
 
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     # 判断用户是否为 POST 提交表单数据
     if request.method == "POST":
         # 将提交的数据赋值到表单实例中
